requirements/local.txt
-r base.txt
django-debug-toolbar==1.7
sqlparse==0.2.3
